splitting and adding js file based on pagetype
