+++ src/testengine.provider.mda/PowerAppsTestEngineMDADashboard.js
@@ -1,0 +1,5 @@
+﻿// WARNING:
+// The JavaScript object model of the page is subject to change. Do not take dependencies on the implementation of methods as they could be updated
+
+class PowerAppsModelDrivenDashboard {
+}
